Fetch and import accounts corrected.
Handle null created user, as deleted users in CC are not passed on.
We provision an unknown user and assign the account to this user (in the fetch).

Alot of logging was added.

There was an error in page, when the cursor is sent the params still need to be sent.